Added ARM CSV

+++ REXCompiler/omptests/rex-simd/benchmark/spmv/rose/rose_spmv_float_sve.c
@@ -105,7 +105,7 @@ int main(int argc,char *argv[])
   double elapsed = read_timer();
   for (row = 0; row < nrows; row++) {
     float sum = 0.0;
-    svbool_t __pg0 = svwhilelt_b32((unsigned long )0,ia[row + 1] - 1);
+    svbool_t __pg0 = svwhilelt_b32((unsigned long )0,(unsigned long)ia[row + 1] - 1);
     svfloat32_t __part0 = svdup_f32(0.00000L);
     for (idx = ia[row]; idx <= ia[row + 1] - 1; idx += svcntw()) {
       svfloat32_t __vec1 = svld1(__pg0,&a[idx]);
